+++ types/util/mongo.d.ts
@@ -1,3 +1,3 @@
 export declare function initMongo(mongoose: unknown, uri: string): Promise<void>;
 
-export declare function destroyMongo(mongoose: unknown): void;
+export declare function destroyMongo(mongoose: unknown): Promise<void>;